+++ rocketgram/api/delete_webhook.py
@@ -4,6 +4,7 @@
 
 
 from dataclasses import dataclass
+from typing import Optional
 
 from .request import Request
 from .utils import BoolResultMixin
@@ -16,3 +17,4 @@ class DeleteWebhook(BoolResultMixin, Request):
     https://core.telegram.org/bots/api#deletewebhook
     """
 
+    drop_pending_updates: Optional[bool] = None